requirements.txt
cartopy
graphviz
ipympl
ipywidgets
jupyterlab
matplotlib
netCDF4
numba
numpy
openpyxl
pandas
pydot
python>=3.9
pyyaml
scikit-learn
scipy
tensorflow>=2.5
tqdm
xarray